feat: Move library code into src/roiti
