+++ src/roiti/gie/__init__.py
@@ -1,0 +1,4 @@
+__version__ = "0.0.1"
+
+from .gie_pandas_client import GiePandasClient
+from .gie_raw_client import GieRawClient
