+++ src/roiti/gie/gie_pandas_client.py
@@ -3,13 +3,13 @@ from typing import Dict, Union, Optional, Any
 
 import pandas as pd
 
-from roiti.gie_raw_client import GieRawClient
-from roiti.mappings.agsi_company import AGSICompany
-from roiti.mappings.agsi_country import AGSICountry
-from roiti.mappings.agsi_facility import AGSIFacility
-from roiti.mappings.alsi_company import ALSICompany
-from roiti.mappings.alsi_country import ALSICountry
-from roiti.mappings.alsi_facility import ALSIFacility
+from .gie_raw_client import GieRawClient
+from .mappings.agsi_company import AGSICompany
+from .mappings.agsi_country import AGSICountry
+from .mappings.agsi_facility import AGSIFacility
+from .mappings.alsi_company import ALSICompany
+from .mappings.alsi_country import ALSICountry
+from .mappings.alsi_facility import ALSIFacility
 
 
 class GiePandasClient(GieRawClient):
